refactor: move GeofencingVertexUpdater to services
