+++ application/src/main/java/org/opentripplanner/service/vehiclerental/street/GeofencingVertexUpdater.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.updater.vehicle_rental;
+package org.opentripplanner.service.vehiclerental.street;
 
 import java.util.Collection;
 import java.util.HashMap;
@@ -15,8 +15,6 @@ import org.locationtech.jts.geom.prep.PreparedGeometry;
 import org.locationtech.jts.geom.prep.PreparedGeometryFactory;
 import org.opentripplanner.framework.geometry.GeometryUtils;
 import org.opentripplanner.service.vehiclerental.model.GeofencingZone;
-import org.opentripplanner.service.vehiclerental.street.BusinessAreaBorder;
-import org.opentripplanner.service.vehiclerental.street.GeofencingZoneExtension;
 import org.opentripplanner.street.model.RentalRestrictionExtension;
 import org.opentripplanner.street.model.edge.Edge;
 import org.opentripplanner.street.model.edge.StreetEdge;
@@ -32,7 +30,7 @@ import org.opentripplanner.street.model.edge.StreetEdge;
  * Perhaps this logic will be replaced with edge splitting where a new vertex is insert right on
  * the border of the zone.
  */
-class GeofencingVertexUpdater {
+public class GeofencingVertexUpdater {
 
   private final Function<Envelope, Collection<Edge>> getEdgesForEnvelope;
 
@@ -44,7 +42,7 @@ class GeofencingVertexUpdater {
    * Applies the restrictions described in the geofencing zones to eges by adding
    * {@link RentalRestrictionExtension} to them.
    */
-  Map<StreetEdge, RentalRestrictionExtension> applyGeofencingZones(
+  public Map<StreetEdge, RentalRestrictionExtension> applyGeofencingZones(
     Collection<GeofencingZone> geofencingZones
   ) {
     var restrictedZones = geofencingZones.stream().filter(GeofencingZone::hasRestriction).toList();
